pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deribit-implied-prob"
version = "0.1.0"
description = "Options-implied probability engine for prediction markets using Deribit data"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.12",
    "pyarrow>=14.0",
    "httpx>=0.26",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "typer>=0.9",
    "rich>=13.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "black>=24.0",
    "mypy>=1.8",
    "jupyter>=1.0",
    "nbformat>=5.9",
]

[project.scripts]
deribit-prob = "src.cli.main:app"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
